moving type declarations to a separate module
