+++ src/dynode/static_value_parameters.py
@@ -1,8 +1,8 @@
 """Provides static parameters to ODEs to solve."""
 
-from . import SEIC_Compartments
 from .abstract_parameters import AbstractParameters
 from .config import Config
+from .typing import SEIC_Compartments
 
 
 class StaticValueParameters(AbstractParameters):
